Update fixed-header.css
Hi! While everything was working flawlessly before, on the latest nova updates the user menu was being pulled off to the right of the screen. I fixed it this way but I am not sure if the problem is here, or on the responsive theme you're using as a base (or on my end, despite re-publishing nova's assets and clearing all cache, it alwasy persisted). Thanks!